% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable}
\alias{variable}
\title{Variable continua de práctica para estratificación}
\format{An object of class \code{numeric} of length 43643.}
\usage{
variable
}
\description{
La variable contiene el índice de marginación urbana para las
AGEBS de México con la información del CENSO 2010. El índice fue
creado por CONAPO a través del método de componentes principales.
}
\keyword{datasets}
